+++ src/Sfa.Tl.ResultsAndCertificationAutomation/Framework/Helpers/ElementHelper.cs
